Refactor test suite to reuse functions used in helium analyser test
